+++ packages/client-reconciliation/src/lib/config.ts
@@ -58,7 +58,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
     language: "typescript",
     openapiDocVersion: "0.0.0",
-    sdkVersion: "1.0.0-alpha.2",
-    genVersion: "2.361.10",
-    userAgent: "speakeasy-sdk/typescript 1.0.0-alpha.2 2.361.10 0.0.0 @formance/sdk-reconciliation",
+    sdkVersion: "1.0.0-alpha.3",
+    genVersion: "2.365.0",
+    userAgent: "speakeasy-sdk/typescript 1.0.0-alpha.3 2.365.0 0.0.0 @formance/sdk-reconciliation",
 } as const;