+++ packages/client-reconciliation/RELEASES.md
@@ -49,3 +49,13 @@ Based on:
 - [typescript v1.0.0-alpha.2] packages/client-reconciliation
 ### Releases
 - [NPM v1.0.0-alpha.2] https://www.npmjs.com/package/@formance/sdk-reconciliation/v/1.0.0-alpha.2 - packages/client-reconciliation
+
+## 2024-07-11 00:16:09
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.331.0 (2.365.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.3] packages/client-reconciliation
+### Releases
+- [NPM v1.0.0-alpha.3] https://www.npmjs.com/package/@formance/sdk-reconciliation/v/1.0.0-alpha.3 - packages/client-reconciliation